use js extension for cache-handler

+++ src/commands/deploy/nextjs/deploy.ts
@@ -439,11 +439,13 @@ function writeNextConfig(cwd: string, region: string) {
     const isCommonJS = existingConfig === "js" || existingConfig === "cjs";
     // Remove .ts extension for TypeScript imports
     const importPath = existingConfig === "ts" ? "./base-next" : `./base-next.${existingConfig}`;
+    const importPathCacheHandler =
+        existingConfig === "ts" ? "./cache-handler.js" : `./cache-handler.${existingConfig}`;
 
     const genezioConfigContent = `
 import userConfig from '${importPath}';
 
-userConfig.cacheHandler = process.env.NODE_ENV === "production" ? "./cache-handler.${existingConfig}" : undefined;
+userConfig.cacheHandler = process.env.NODE_ENV === "production" ? "${importPathCacheHandler}" : undefined;
 userConfig.cacheMaxMemorySize = 0;
 
 ${isCommonJS ? "module.exports = userConfig;" : "export default userConfig;"}
@@ -451,10 +453,14 @@ ${isCommonJS ? "module.exports = userConfig;" : "export default userConfig;"}
 
     fs.writeFileSync(genezioConfigPath, genezioConfigContent);
 
-    fs.writeFileSync(
-        path.join(cwd, `cache-handler.${existingConfig}`),
-        getCacheHandlerContent(existingConfig as "js" | "ts" | "mjs", region),
-    );
+    if (existingConfig === "ts") {
+        fs.writeFileSync(path.join(cwd, `cache-handler.js`), getCacheHandlerContent("js", region));
+    } else {
+        fs.writeFileSync(
+            path.join(cwd, `cache-handler.${existingConfig}`),
+            getCacheHandlerContent(existingConfig as "js" | "ts" | "mjs", region),
+        );
+    }
 }
 
 function determineFileExtension(cwd: string): "js" | "mjs" | "ts" {
